requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
arrow
async-timeout
attrs
beautifulsoup4
bitsandbytes
blinker
bs4
cachetools
certifi
cffi
charset-normalizer
circuitbreaker
click
cohere
cryptography
dataclasses-json
et_xmlfile
exceptiongroup
fastavro
filelock
frozenlist
fsspec
gitdb
GitPython
greenlet
grpcio
grpcio-tools
h11
h2
hpack
httpcore
httpx
httpx-sse
huggingface-hub
hyperframe
idna
Jinja2
jmespath
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-cohere
langchain-community
langchain-core
langchain-experimental
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
mpmath
multidict
mypy-extensions
narwhals
networkx
numpy
openpyxl
oracledb
orjson
outcome
packaging
pandas
parameterized
pdf2image
pdfminer.six
pdfplumber
pillow
playwright
portalocker
prompt-toolkit
propcache
protobuf
psutil
pyarrow
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
pyee
Pygments
pyOpenSSL
pypdf
PyPDF2
pycryptodome
pypdfium2
PySocks
pytesseract
python-dateutil
python-dotenv
pytz
PyYAML
qdrant-client
referencing
regex
requests
requests-toolbelt
rich
rpds-py
safetensors
scipy
selenium
six
smmap
sniffio
sortedcontainers
soupsieve
SQLAlchemy
streamlit
sympy
tabulate
tenacity
terminaltables
tokenizers
toml
torch
tornado
tqdm
transformers @ git+https://github.com/huggingface/transformers@3a8eb74668e9c2cc563b2f5c62fac174797063e0
trio
trio-websocket
types-python-dateutil
types-requests
typing-inspect
typing_extensions
tzdata
urllib3
watchdog
wcwidth
websocket-client
wsproto
yarl
youtube-transcript-api